Chapter 10, 11

+++ classes/Ijdb/Controllers/Joke.php
@@ -2,15 +2,18 @@
     namespace Ijdb\Controllers;
 
     use \Hanbit\DatabaseTable;
+    use \Hanbit\Authentication;
 
     class Joke {
 
         private $jokesTable;
         private $authorsTable;
+        private $authentication;
 
-        public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable) {
+        public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable, Authentication $authentication) {
             $this -> jokesTable = $jokesTable;
             $this -> authorsTable = $authorsTable;
+            $this -> authentication = $authentication;
         }
 
         public function home() {
@@ -32,7 +35,8 @@
                     'joketext' => $joke['joketext'],
                     'jokedate' => $joke['jokedate'],
                     'name' => $author['name'],
-                    'email' => $author['email']
+                    'email' => $author['email'],
+                    'authorId' => $author['id']
                 ];
             }
 
@@ -40,20 +44,33 @@
 
             $totalJokes = $this -> jokesTable -> total();
 
-            return ['template' => 'jokes2.html.php',
-                    'title' => $title,
-                    'variables' => [
-                        'totalJokes' => $totalJokes,
-                        'jokes' => $jokes
-                    ]
-                ];
+            $author = $this -> authentication -> getUser();
+
+            return ['template' => 'jokes.html.php',
+                'title' => $title,
+                'variables' => [
+                    'totalJokes' => $totalJokes,
+                    'jokes' => $jokes,
+                    'userId' => $author['id'] ?? null
+                ]
+            ];
         }
 
         public function saveEdit() {
+            $author = $this -> authentication -> getUser();
+
+            if (isset($_GET['id'])) {
+                $joke = $this -> jokesTable -> findById($_GET['id']);
+
+                if ($joke['authorId'] != $author['id']) {
+                    return;
+                }
+            }
+
             $joke = $_POST['joke'];
-    
-            $joke['authorId'] = 1;
+
             $joke['jokedate'] = new \DateTime();
+            $joke['authorId'] = $author['id'];
 
             $this -> jokesTable -> save($joke);
 
@@ -61,6 +78,14 @@
         }
 
         public function edit() {
+            if (!$this -> authentication -> isLoggedIn()) {
+                return ['template' => 'loginerror.html.php',
+                    'title' => '이 페이지를 볼 수 있는 권한이 없습니다.'
+                ];
+            }
+
+            $author = $this -> authentication -> getUser();
+
             if (isset($_GET['id'])) {
                 $joke = $this -> jokesTable -> findById($_GET['id']);
             }
@@ -70,12 +95,21 @@
             return ['template' => 'editjoke.html.php',
                 'title' => $title,
                 'variables' => [
-                    'joke' => $joke ?? null
+                    'joke' => $joke ?? null,
+                    'userId' => $author['id'] ?? null
                 ]
             ];
         }
 
         public function delete() {
+            $author = $this -> authentication -> getUser();
+
+            $joke = $this -> jokesTable -> findById($_GET['id']);
+
+            if ($joke['authorId'] != $author['id']) {
+                return;
+            }
+
             $this -> jokesTable -> delete($_POST['id']);
 
             header('location: /joke/list');
